identity-service/requirements.txt
fastapi
uvicorn[standard]
PyJWT
passlib==1.7.4
bcrypt==3.2.2